requirements.txt
cython==0.29.26
libscrc==1.7
numpy==1.21.4
pandas==1.3.5
pyserial==3.5
termcolor==1.1.0
